requirements.txt
boto3==1.36.12
chromadb==0.6.3
Flask==3.1.0
flask_cors==5.0.0
gensim==4.3.3
hdbscan==0.8.40
joblib==1.4.2
matplotlib==3.8.4
nltk==3.8.1
numpy
openai==1.61.0
pandas==2.2.3
pinecone==5.4.2
praw==7.8.1
pytest==7.4.4
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.4.2
tensorflow==2.18.0
tqdm==4.66.4
umap==0.1.1
umap-learn
gunicorn
python-dotenv